v4 with attempted optimization akin to web components
The optimization would, in theory, allow each function component to act as its own tree. While it works on the surface, there are major issues going on with the attributes when rendering the parent and sub tree together. Hooks are also a major pain to implement.

I will have to look at ways to review the rendering and mounting process to reconcile created tree with their parent trees. If a child tree is rendered in isolation, it works, but if a parent tree render a child tree, then everything breaks.

This was a cool experiment, but it needs more investigation.