Improve the context API a lot and fix the type weirdness
The new context API should not be a bit more intuitive to React users and allow a lot more customization without the headaches of the old API. Also tried a quick optimization with the rendering process by deferring writes until reads have all been completed.

This has yet to be tested.