Add docs and fix security issues in dependencies
The library was essentially abandoned and had been for multiple years. Though I do plan on working on it in the future, this will likely not happen soon and leaving the library without any README felt wrong. As part of a cleanup of my random projects, I added more details as to why this library came to be, what it attempts to achieve, and what came out of those experiments.